+++ lib/ecto/adapters/clickhouse/connection.ex
@@ -83,39 +83,29 @@ defmodule Ecto.Adapters.ClickHouse.Connection do
     ]
   end
 
-  @dialyzer {:no_return, update_all: 1, update_all: 2}
   @impl true
-  def update_all(query, _prefix \\ nil) do
-    raise Ecto.QueryError,
-      query: query,
-      message:
-        "ClickHouse does not support UPDATE statements -- use ALTER TABLE ... UPDATE instead"
+  # https://clickhouse.com/docs/sql-reference/statements/update
+  def update_all(query, params \\ []) do
+    validate_update_all!(query, "UPDATE")
+
+    %{sources: sources} = query
+    {table, _schema, prefix} = elem(sources, 0)
+    fields = update_fields(query, sources, params)
+
+    where =
+      case query.wheres do
+        [] -> " WHERE 1"
+        _ -> where(query, {{nil, nil, nil}, []}, params)
+      end
+
+    ["UPDATE ", quote_table(prefix, table), " SET ", fields, where]
   end
 
   # https://clickhouse.com/docs/en/sql-reference/statements/alter/update
   # https://clickhouse.com/docs/en/guides/developer/mutations#updating-data
   def alter_update_all(query, params \\ []) do
     # TODO link to https://clickhouse.com/blog/handling-updates-and-deletes-in-clickhouse#updating-and-deleting-using-joins
-    unless query.joins == [] do
-      raise Ecto.QueryError,
-        query: query,
-        message:
-          "Ecto.Adapters.ClickHouse does not support JOIN in ALTER TABLE ... UPDATE statements"
-    end
-
-    if query.select do
-      raise Ecto.QueryError,
-        query: query,
-        message:
-          "Ecto.Adapters.ClickHouse does not support RETURNING in ALTER TABLE ... UPDATE statements"
-    end
-
-    if query.with_ctes do
-      raise Ecto.QueryError,
-        query: query,
-        message:
-          "Ecto.Adapters.ClickHouse does not support CTEs in ALTER TABLE ... UPDATE statements"
-    end
+    validate_update_all!(query, "ALTER TABLE ... UPDATE")
 
     %{sources: sources} = query
     {table, _schema, prefix} = elem(sources, 0)
@@ -128,6 +118,26 @@ defmodule Ecto.Adapters.ClickHouse.Connection do
       end
 
     ["ALTER TABLE ", quote_table(prefix, table), " UPDATE ", fields, where]
+  end
+
+  defp validate_update_all!(query, statement) do
+    unless query.joins == [] do
+      raise Ecto.QueryError,
+        query: query,
+        message: "Ecto.Adapters.ClickHouse does not support JOIN in #{statement} statements"
+    end
+
+    if query.select do
+      raise Ecto.QueryError,
+        query: query,
+        message: "Ecto.Adapters.ClickHouse does not support RETURNING in #{statement} statements"
+    end
+
+    if query.with_ctes do
+      raise Ecto.QueryError,
+        query: query,
+        message: "Ecto.Adapters.ClickHouse does not support CTEs in #{statement} statements"
+    end
   end
 
   defp update_fields(%{updates: updates} = query, sources, params) do
@@ -242,9 +252,32 @@ defmodule Ecto.Adapters.ClickHouse.Connection do
   end
 
   @impl true
-  def update(_prefix, _table, _fields, _filters, _returning) do
-    raise ArgumentError,
-          "ClickHouse does not support UPDATE statements -- use ALTER TABLE ... UPDATE instead"
+  def update(prefix, table, fields, filters, returning) do
+    unless returning == [] do
+      raise ArgumentError, "ClickHouse does not support RETURNING on UPDATE statements"
+    end
+
+    field_count = length(fields)
+
+    fields =
+      fields
+      |> Enum.with_index()
+      |> intersperse_map(?,, fn {{field, value}, idx} ->
+        [quote_name(field), ?=, build_param(idx, value)]
+      end)
+
+    filters =
+      filters
+      |> Enum.with_index(field_count)
+      |> intersperse_map(" AND ", fn
+        {{field, nil}, _idx} ->
+          ["isNull(", quote_name(field), ?)]
+
+        {{field, value}, idx} ->
+          [quote_name(field), ?=, build_param(idx, value)]
+      end)
+
+    ["UPDATE ", quote_table(prefix, table), " SET ", fields, " WHERE ", filters]
   end
 
   @impl true
@@ -1228,6 +1261,7 @@ defmodule Ecto.Adapters.ClickHouse.Connection do
 
   defp param_type(f) when is_float(f), do: "Float64"
   defp param_type(b) when is_boolean(b), do: "Bool"
+  defp param_type(nil), do: "Nullable(Nothing)"
 
   # TODO DateTime timezone?
   defp param_type(%s{microsecond: microsecond}) when s in [NaiveDateTime, DateTime] do